Crate autoscaler1_beta2

Features

Handle the following Resources with ease from the central hub ...

autoscalers (delete, get, insert, list, patch and update)

zone operations (delete, get and list)

zones (list)

Everything else about the autoscaler API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.autoscalers().list(...).doit()
let r = hub.autoscalers().update(...).doit()
let r = hub.autoscalers().patch(...).doit()
let r = hub.autoscalers().delete(...).doit()
let r = hub.autoscalers().insert(...).doit()
let r = hub.autoscalers().get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate autoscaler1_beta2;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and `client_secret`, 
// among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about what's going on
// You probably want to bring in your own `TokenStorage` to persist tokens and retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AutoscalerHub::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.autoscalers().list("project", "zone")
             .page_token("accusam")
             .max_results(93)
             .filter("justo")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(value) => println!("Result Value: {:?}", value),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

Autoscaler
Cloud Autoscaler resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (none)

update (request)

patch (request)

delete (none)

insert (request)

get (response)

AutoscalerDeleteMethodBuilder
Deletes the specified Autoscaler resource.

AutoscalerGetMethodBuilder
Gets the specified Autoscaler resource.

AutoscalerHub
Central instance to access all AutoscalerHub related resource activities

AutoscalerInsertMethodBuilder
Adds new Autoscaler resource.

AutoscalerListMethodBuilder
Lists all Autoscaler resources in this zone.

AutoscalerListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

AutoscalerMethodsBuilder
A builder providing access to all methods supported on autoscaler resources. It is not used directly, but through the AutoscalerHub hub.

AutoscalerPatchMethodBuilder
Update the entire content of the Autoscaler resource. This method supports patch semantics.

AutoscalerUpdateMethodBuilder
Update the entire content of the Autoscaler resource.

AutoscalingPolicy
Cloud Autoscaler policy.

This type is not used in any activity, and only used as part of another schema.

AutoscalingPolicyCpuUtilization
CPU utilization policy.

This type is not used in any activity, and only used as part of another schema.

AutoscalingPolicyCustomMetricUtilization
Custom utilization metric policy.

This type is not used in any activity, and only used as part of another schema.

AutoscalingPolicyLoadBalancingUtilization
Load balancing utilization policy.

This type is not used in any activity, and only used as part of another schema.

DeprecationStatus
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Operation
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update (response)

delete (response)

insert (response)

get (response)

patch (response)

OperationError
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

OperationErrorErrors
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

OperationList
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

OperationWarnings
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

OperationWarningsData
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Zone
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (none)

ZoneList
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

ZoneListMethodBuilder

ZoneMaintenanceWindows
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

ZoneMethodsBuilder
A builder providing access to all methods supported on zone resources. It is not used directly, but through the AutoscalerHub hub.

ZoneOperationDeleteMethodBuilder
Deletes the specified zone-specific operation resource.

ZoneOperationGetMethodBuilder
Retrieves the specified zone-specific operation resource.

ZoneOperationListMethodBuilder
Retrieves the list of operation resources contained within the specified zone.

ZoneOperationMethodsBuilder
A builder providing access to all methods supported on zoneOperation resources. It is not used directly, but through the AutoscalerHub hub.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.